# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Resource-Service is service for storing kubernetes resources in Mongo DB.",
    "title": "Resource-Service",
    "version": "0.0.1"
  },
  "paths": {
...
    "/namespaces/{namespace}/services": {
      "get": {
        "tags": [
          "Service"
        ],
        "summary": "Get services list.",
        "operationId": "GetServicesListHandler",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "X-User-Namespace",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "services list",
            "schema": {
              "$ref": "#/definitions/ServicesResponse"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "public"
      },
      "post": {
        "tags": [
          "Service"
        ],
        "summary": "Create service.",
        "operationId": "CreateServiceHandler",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "X-User-Namespace",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Service"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "service created",
            "schema": {
              "$ref": "#/definitions/ResourceService"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "public"
      },
      "delete": {
        "tags": [
          "Service"
        ],
        "summary": "Delete service.",
        "operationId": "DeleteAllServicesHandler",
        "parameters": [
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "all services in namespace deleted"
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "private"
      }
    },
    "/namespaces/{namespace}/services/{service}": {
      "get": {
        "tags": [
          "Service"
        ],
        "summary": "Get services list.",
        "operationId": "GetServiceHandler",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "X-User-Namespace",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "service",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "service",
            "schema": {
              "$ref": "#/definitions/ResourceService"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "public"
      },
      "put": {
        "tags": [
          "Service"
        ],
        "summary": "Update service.",
        "operationId": "UpdateServiceHandler",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "X-User-Namespace",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "service",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Service"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "service updated",
            "schema": {
              "$ref": "#/definitions/ResourceService"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "public"
      },
      "delete": {
        "tags": [
          "Service"
        ],
        "summary": "Delete service.",
        "operationId": "DeleteServiceHandler",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "X-User-Namespace",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "service",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "service deleted"
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "public"
      }
    },
...
    "/namespaces/{namespace}/solutions/{solution}/services": {
      "delete": {
        "tags": [
          "Service"
        ],
        "summary": "Delete all solution services.",
        "operationId": "DeleteAllSolutionServicesHandler",
        "parameters": [
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "solution",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "all solution services deleted"
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "private"
      }
    },
...
  },
  "definitions": {
...
    "ListService": {
      "description": "ListService -- services list",
      "type": "array",
      "items": {
        "$ref": "#/definitions/ResourceService"
      },
      "x-go-package": "git.containerum.net/ch/resource-service/pkg/models/service"
    },
...
    "Protocol": {
      "description": "represents port protocol type (TCP or UDP)",
      "type": "string",
      "x-go-package": "git.containerum.net/ch/resource-service/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "ResourceService": {
      "description": "ResourceService --  model for service for resource-service db",
      "type": "object",
      "required": [
        "name",
        "ports"
      ],
      "properties": {
        "_id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "created_at": {
          "description": "creation date in RFC3339 format",
          "type": "string",
          "x-go-name": "CreatedAt"
        },
        "deleted": {
          "type": "boolean",
          "x-go-name": "Deleted"
        },
        "deleted_at": {
          "description": "delete date in RFC3339 format",
          "type": "string",
          "x-go-name": "DeletedAt"
        },
        "deploy": {
          "type": "string",
          "x-go-name": "Deploy"
        },
        "domain": {
          "type": "string",
          "x-go-name": "Domain"
        },
        "ips": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "IPs"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "namespaceid": {
          "type": "string",
          "x-go-name": "NamespaceID"
        },
        "owner": {
          "type": "string",
          "x-go-name": "Owner"
        },
        "ports": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ServicePort"
          },
          "x-go-name": "Ports"
        },
        "solution_id": {
          "description": "Solution ID (only if service is part of solution)",
          "type": "string",
          "x-go-name": "SolutionID"
        },
        "type": {
          "$ref": "#/definitions/Type"
        }
      },
      "x-go-package": "git.containerum.net/ch/resource-service/pkg/models/service"
    },
...
    "Service": {
      "description": "represents service",
      "type": "object",
      "required": [
        "name",
        "ports"
      ],
      "properties": {
        "created_at": {
          "description": "creation date in RFC3339 format",
          "type": "string",
          "x-go-name": "CreatedAt"
        },
        "deleted_at": {
          "description": "delete date in RFC3339 format",
          "type": "string",
          "x-go-name": "DeletedAt"
        },
        "deploy": {
          "type": "string",
          "x-go-name": "Deploy"
        },
        "domain": {
          "type": "string",
          "x-go-name": "Domain"
        },
        "ips": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "IPs"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "owner": {
          "type": "string",
          "x-go-name": "Owner"
        },
        "ports": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ServicePort"
          },
          "x-go-name": "Ports"
        },
        "solution_id": {
          "description": "Solution ID (only if service is part of solution)",
          "type": "string",
          "x-go-name": "SolutionID"
        }
      },
      "x-go-package": "git.containerum.net/ch/resource-service/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "ServicePort": {
      "description": "represent service port",
      "type": "object",
      "required": [
        "name",
        "target_port",
        "protocol"
      ],
      "properties": {
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "port": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Port"
        },
        "protocol": {
          "$ref": "#/definitions/Protocol"
        },
        "target_port": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "TargetPort"
        }
      },
      "x-go-package": "git.containerum.net/ch/resource-service/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "ServicesResponse": {
      "description": "ServicesResponse -- ingresses response",
      "type": "object",
      "properties": {
        "services": {
          "$ref": "#/definitions/ListService"
        }
      },
      "x-go-package": "git.containerum.net/ch/resource-service/pkg/models/service"
    },
...
    "Type": {
      "type": "string",
      "x-go-package": "git.containerum.net/ch/resource-service/pkg/models/service"
    },
...
    "errOAIGen": {
      "description": "Err -- standard serializable API error\nMessage -- constant error message:\n+ \"invalid username\"\n+ \"quota exceeded\"\n+ \"validation error\"\n...etc...\nID -- unique error identification code\nDetails -- optional context error messages kinda\n+ \"field 'Replicas' must be non-zero value\"\n+ \"not enough tights to feed gopher\"\n+ \"resource 'God' does't exist\"\nFields -- optional set of key-value pairs",
      "type": "object",
      "properties": {
        "details": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Details"
        },
        "fields": {
          "description": "Fields -- represents additional key-value fields of error",
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-package": "github.com/containerum/cherry"
        },
        "id": {
          "$ref": "#/definitions/errOaiGenId"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "status_http": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StatusHTTP"
        }
      },
      "x-go-package": "github.com/containerum/cherry"
    },
    "errOaiGenId": {
      "description": "ErrID -- represents unique error ID",
      "type": "object",
      "properties": {
        "kind": {
          "description": "ErrKind -- represents kind of error",
          "type": "integer",
          "format": "uint64",
          "x-go-package": "github.com/containerum/cherry"
        },
        "sid": {
          "description": "ErrSID -- represents service ID of error",
          "type": "string",
          "x-go-package": "github.com/containerum/cherry"
        }
      },
      "x-go-gen-location": "models",
      "x-go-package": "github.com/containerum/cherry"
    }
  },
...
}